Finished Add / Edit system

+++ Sporadic/Models/Activity.swift
@@ -10,8 +10,8 @@ import Foundation
 struct Activity: Identifiable, Codable, Equatable {
     var id: Int = 0
     var name: String = "Run"
-    var minValue: Float = 1
-    var maxValue: Float = 5
-    var total: Float = 0
+    var minValue: Double = 1
+    var maxValue: Double = 5
+    var total: Double = 0
     var isEnabled: Bool = false
 }
